% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build a compartment model}
\usage{
build_model(...)
}
\arguments{
\item{...}{Compartments supplied individually, or a single list of
compartments.}
}
\value{
An object of class `"cm_model"` containing named compartments and
  an initially `NULL` flow function.
}
\description{
Construct a model object from one or more compartment objects.
}
\examples{
s <- new_compartment("S", 999)
i <- new_compartment("I", 1)
m <- build_model(s, i)
m

}
