% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{Logistic Transition Function}
\usage{
delta(x, scale, start)
}
\arguments{
\item{x}{Numeric vector representing time or another continuous variable.}

\item{scale}{Scaling factor controlling the rate of transition.}

\item{start}{The time point at which the transition begins.}
}
\value{
A numeric vector representing the transition effect, ranging between 0 and 1.
}
\description{
This function models a smooth logistic transition using a scaled hyperbolic tangent function.
}
\details{
The function applies a hyperbolic tangent transformation to model a smooth transition. The output is shifted and scaled to range between 0 and 1.
}
\examples{
x <- seq(0, 100, by = 1)
scale <- 10
start <- 50
delta_effect <- delta(x, scale, start)
plot(x, delta_effect, type = "l", main = "Logistic Transition Over Time")

}
