% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epi_MPC_run_comp_model.R
\name{Epi_MPC_run_comp_model}
\alias{Epi_MPC_run_comp_model}
\title{Runs MPC while stepping a custom compartmental model
Requires:
  - model: built via build_model()
  - step_model(), state(), flow_SIRDS() that attaches events via attr(delta, "events")
  - episimdata: has columns for compartments and events
  - R_estimator(): function to estimate R
  - action space
  - updatepars: the function defining how interverntions affect transmission
  - disease parameters}
\usage{
Epi_MPC_run_comp_model(
  episimdata,
  episettings,
  epi_par,
  noise_par,
  actions,
  pred_days,
  n_ens = 100,
  start_day = 1,
  ndays = nrow(episimdata),
  R_est_wind = 5,
  pathogen = 1,
  susceptibles = 0,
  delay = 0,
  ur = 0,
  r_dir = 1,
  N = sim_settings$N
)
}
\description{
Runs MPC while stepping a custom compartmental model
Requires:
  - model: built via build_model()
  - step_model(), state(), flow_SIRDS() that attaches events via attr(delta, "events")
  - episimdata: has columns for compartments and events
  - R_estimator(): function to estimate R
  - action space
  - updatepars: the function defining how interverntions affect transmission
  - disease parameters
}
