% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epi_pred.R
\name{Epi_pred}
\alias{Epi_pred}
\title{Simulate an Epidemic Without Control ('Open-Loop') -- use when death data is not available}
\usage{
Epi_pred(
  episimdata,
  epi_par,
  noise_par,
  actions,
  pathogen,
  pred_days,
  kk,
  jj,
  N,
  ndays = nrow(episimdata),
  pred_susceptibles = 0,
  gamma = 0.95
)
}
\arguments{
\item{episimdata}{A data frame containing epidemic simulation data.}

\item{epi_par}{A matrix of epidemiological parameters, with rows corresponding to different pathogens.}

\item{noise_par}{A parameter related to stochastic noise in the epidemic simulation (not used in the function).}

\item{actions}{A matrix of control actions, where the second column modifies the estimated reproduction number.}

\item{pathogen}{A character or numeric index specifying the pathogen of interest in `epi_par`.}

\item{pred_days}{The number of days to predict forward.}

\item{kk}{The starting day of prediction.}

\item{jj}{The index of the action scenario being simulated.}

\item{N}{The total population size.}

\item{ndays}{The total number of days in the epidemic simulation (default: `nrow(episimdata)`).}

\item{pred_susceptibles}{Logical (0 or 1), indicating whether susceptible population dynamics should be accounted for.}

\item{gamma}{The discount factor for future rewards (default: 0.95).}
}
\value{
The expected reward (`Exp_rew`) over the simulation period.
}
\description{
This function simulates the progression of an epidemic without intervention, using predefined parameters.
}
\details{
This function updates the effective reproduction number (`Re`), computes the expected number of new cases (`C`), and applies a discount factor to rewards computed using `reward_fun`. The function assumes that cases follow a Poisson process and uses a gamma-distributed generation time to estimate the infection dynamics.
}
