% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epi_pred_wd.R
\name{Epi_pred_wd}
\alias{Epi_pred_wd}
\title{Simulate Epidemic Dynamics and Compute Expected Reward}
\usage{
Epi_pred_wd(
  episimdata,
  epi_par,
  noise_par,
  actions,
  pathogen,
  pred_days,
  r_dir,
  kk,
  jj,
  N,
  ndays = nrow(episimdata),
  pred_susceptibles = 0,
  gamma = 0.95
)
}
\arguments{
\item{episimdata}{A data frame containing simulation data. It should include columns such as
\code{"C"} (cases), \code{"I"} (infected individuals), \code{"Re"} (effective reproduction number),
\code{"S"} (susceptible individuals), \code{"Deaths"}, and \code{"Lambda"}.}

\item{epi_par}{A data frame containing epidemiological parameters for various pathogens. It should
have the following columns: \code{"R0"} (basic reproduction number), \code{"gen_time"}
(generation time), \code{"gen_time_var"} (variance of generation time), \code{"CFR"}
(case fatality rate), \code{"mortality_mean"}, and \code{"mortality_var"}.}

\item{noise_par}{A placeholder for noise parameters. Not used in projections.}

\item{actions}{A data frame containing control actions. Column 2 is expected to modify the effective
reproduction number (\code{"Re"}).}

\item{pathogen}{A string specifying the pathogen name to extract corresponding epidemiological parameters.}

\item{pred_days}{An integer specifying the number of days to predict ahead.}

\item{r_dir}{An integer specifying the reproduction number adjustments:
\itemize{
  \item \code{1} for direct \code{Re}.
  \item \code{2} for logistic adjustments.
  \item \code{0} for using the generation time distribution.
}}

\item{kk}{An integer indicating the starting day for prediction within the simulation.}

\item{jj}{An integer specifying the row index in \code{actions} to use for control effects.}

\item{N}{A numeric value representing the total population size.}

\item{ndays}{An integer specifying the total number of days in the simulation. Defaults to the number
of rows in \code{episimdata}.}

\item{pred_susceptibles}{A binary (0 or 1) indicating whether to update the number of susceptibles
during the simulation. Defaults to \code{0}.}

\item{gamma}{A numeric value between 0 and 1 representing the discount factor for future rewards. Defaults to \code{0.95}. Smaller values will prioritise
immediate rewards over longer term rewards.}
}
\value{
A numeric value representing the expected discounted reward over the prediction window.
}
\description{
This function simulates an epidemic using pre-defined parameters
and computes the expected reward over a specified projection window.
}
\details{
The function simulates the epidemic using specified parameters and computes rewards for each
day within the prediction window. It supports dynamic updates for cases, reproduction numbers,
and deaths based on a Poisson process and delayed distributions for secondary cases and mortality.
Rewards are calculated using the \code{\link{reward_fun_wd}} function and are discounted
exponentially using the discount factor \code{gamma}.
}
\examples{
# Example epidemiological data
episimdata <- data.frame(R0est = c(1.5, 1.6), C = c(0, 10), Re = c(NA, NA), S = c(1000, 990), Deaths = c(0, 1))
epi_par <- data.frame(
  R0 = c(2.5), gen_time = c(5), gen_time_var = c(1),
  CFR = c(0.02), mortality_mean = c(14), mortality_var = c(2)
)
actions <- data.frame(action_effect = c(0.9, 0.8))
Epi_pred_wd(
  episimdata = episimdata, epi_par = epi_par, noise_par = NULL,
  actions = actions, pathogen = "pathogen1", pred_days = 10,
  r_dir = 1, kk = 2, jj = 1, N = 1000, ndays = 20
)

}
