% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicontrol.R
\name{epicontrol}
\alias{epicontrol}
\title{Run Epidemiological Simulations with Control}
\usage{
epicontrol(episim_data_ens, episettings)
}
\arguments{
\item{episim_data_ens}{A list of data frames containing initial epidemiological simulation data for each ensemble member.}

\item{episettings}{A list of settings controlling the simulation, including:
- `sim_function`: Function to run a single epidemiological simulation.
- `reward_function`: Function to calculate rewards for the control process.
- `R_estimator`: Function to estimate the reproduction number.
- `noise_par`: Parameters related to noise in the simulation.
- `epi_par`: Epidemiological parameters.
- `actions`: Possible intervention actions.
- `sim_settings`: A list of simulation-specific settings. Please refer to the sim_function's documentation for the list of required parameters
- `parallel`: Logical, whether to run simulations in parallel.
- `cl`: Cluster object for parallel computation if `parallel = TRUE`.}
}
\value{
A list of simulation results for each ensemble member.
}
\description{
This function runs multiple ensemble simulations for epidemiological modelling,
incorporating noise parameters, estimated reproduction numbers, and intervention actions.
It supports parallel computation for efficiency.
}
