% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_compartment.R
\name{new_compartment}
\alias{new_compartment}
\title{Create a compartment}
\usage{
new_compartment(name, prevalence)
}
\arguments{
\item{name}{Character scalar. Name of the compartment.}

\item{prevalence}{Numeric scalar. Initial prevalence of the compartment.}
}
\value{
An object of class `"compartment"`.
}
\description{
Construct a compartment object with a name, prevalence, and a placeholder
for change.
}
\examples{
s <- new_compartment("S", 999)
i <- new_compartment("I", 1)
s

}
