% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_estim.R
\name{R_estim}
\alias{R_estim}
\title{Estimate Reproduction Number (R) and the Reduction Coefficient}
\usage{
R_estim(episimdata, Ygen, ii, R_est_wind = 5, r_dir = 0)
}
\arguments{
\item{episimdata}{A data frame containing epidemic simulation data with columns `C`, `Lambda_C`, and `R_coeff` containing, cases, total infectiousness (computed form cases), and the reduction in R by policy, respectively.}

\item{Ygen}{A numeric vector representing the generation time distribution.}

\item{ii}{An integer specifying the current time step in the simulation.}

\item{R_est_wind}{An integer (default = 5) defining the window size for R estimation.}

\item{r_dir}{An integer specifying the reproduction number adjustments.}
}
\value{
A list containing:
  - `R_est`: Estimated reproduction number.
  - `R_coeff_tmp`: Estimated reduction coefficient for `R0`.
}
\description{
This function estimates the reproduction number (`R_est`) and an associated coefficient (`R_coeff_tmp`)
based on epidemic ata.
}
