% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_flows.R
\name{set_flows}
\alias{set_flows}
\title{Set the model flow function}
\usage{
set_flows(model, flow_fun)
}
\arguments{
\item{model}{A `"cm_model"` object.}

\item{flow_fun}{A function describing model flows. It should accept
arguments `t`, `state`, `pars`, and `dt`, and return a named vector of
compartment changes. It may also include a named `"events"` attribute.}
}
\value{
The updated `"cm_model"` object.
}
\description{
Attach a user-defined flow function to a compartment model.
}
\examples{
s <- new_compartment("S", 999)
i <- new_compartment("I", 1)
m <- build_model(s, i)

flow_fun <- function(t, state, pars, dt) {
  c(S = -1, I = 1)
}

m <- set_flows(m, flow_fun)

}
