% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state}
\alias{state}
\title{Extract the current state vector}
\usage{
state(model)
}
\arguments{
\item{model}{A `"cm_model"` object.}
}
\value{
A named numeric vector of compartment prevalences.
}
\description{
Return the current compartment prevalences as a named numeric vector.
}
\examples{
s <- new_compartment("S", 999)
i <- new_compartment("I", 1)
m <- build_model(s, i)
state(m)

}
