% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vac.R
\name{vac}
\alias{vac}
\title{Vaccination Effect Over Time}
\usage{
vac(x, maxv, scale, start)
}
\arguments{
\item{x}{Numeric vector representing time or another continuous variable.}

\item{maxv}{Maximum vaccination effect or uptake level.}

\item{scale}{Scaling factor controlling the rate of change.}

\item{start}{The time point at which vaccination starts to take effect.}
}
\value{
A numeric vector representing the vaccination effect, ensuring non-negative values.
}
\description{
This function models the effect of vaccination over time using a scaled hyperbolic tangent function.
}
\details{
The function applies a hyperbolic tangent transformation to model a smooth transition of vaccination uptake. The result is capped at a minimum of zero using `pmax()` to prevent negative values.
}
\examples{
x <- seq(0, 100, by = 1)
maxv <- 0.8
scale <- 10
start <- 50
vac_effect <- vac(x, maxv, scale, start)
plot(x, vac_effect, type = "l", main = "Vaccination Effect Over Time")

}
